refactor(sdk): extract user decrypt response processing into client-core
Move UserDecryptionResponseBuilder, deserializer, and ResponseConfig
from rust-sdk to client-core so UniFFI and future platform targets
can reuse the KMS decryption logic without depending on gateway-sdk.

Adds kms, kms-grpc, bc2wrap (default-features=false) and ml-kem
as client-core dependencies using the WASM-compatible code paths only.
