+++ sdk/client-core/src/decryption/user/deserializer.rs
@@ -1,4 +1,4 @@
-use crate::{FhevmError, Result};
+use crate::{ClientCoreError, Result};
 use kms_grpc::kms::v1::{UserDecryptionResponse, UserDecryptionResponsePayload};
 use serde::Deserialize;
 use tracing::debug;
@@ -39,7 +39,7 @@ impl UserDecryptionDeserializer {
 
         // Step 1: Parse JSON to hex intermediate type
         let hex_responses: JsonResponse = serde_json::from_str(json_str)
-            .map_err(|e| FhevmError::DecryptionError(format!("JSON parse error: {e}")))?;
+            .map_err(|e| ClientCoreError::DecryptionError(format!("JSON parse error: {e}")))?;
 
         debug!(
             "Found {} responses to process",
@@ -93,7 +93,7 @@ fn decode_hex_field(hex_str: &str, field_name: &str) -> Result<Vec<u8>> {
     let cleaned = hex_str.trim_start_matches("0x");
 
     hex::decode(cleaned)
-        .map_err(|e| FhevmError::DecryptionError(format!("Invalid {field_name} hex: {e}")))
+        .map_err(|e| ClientCoreError::DecryptionError(format!("Invalid {field_name} hex: {e}")))
 }
 
 /// Deserialize payload bytes to UserDecryptionResponsePayload
@@ -101,7 +101,7 @@ fn deserialize_payload(buf: &[u8]) -> Result<UserDecryptionResponsePayload> {
     debug!("Deserializing payload of {} bytes", buf.len());
 
     bc2wrap::deserialize_safe(buf)
-        .map_err(|e| FhevmError::DecryptionError(format!("Bincode deserialize error: {e}")))
+        .map_err(|e| ClientCoreError::DecryptionError(format!("Bincode deserialize error: {e}")))
 }
 
 #[cfg(test)]
@@ -130,7 +130,7 @@ mod tests {
         assert!(responses[0].payload.is_some());
 
         if let Some(payload) = &responses[0].payload {
-            println!("✅ Payload deserialized successfully!");
+            println!("Payload deserialized successfully!");
             println!(
                 "   Verification key: {} bytes",
                 payload.verification_key.len()
