+++ sdk/client-core/src/decryption/user/types.rs
@@ -24,3 +24,37 @@ pub struct DecryptedValue {
     pub value: Vec<u8>,
     pub fhe_type: i32,
 }
+
+/// Configuration for response processing (builder internal state).
+#[derive(Debug)]
+pub(crate) struct ResponseConfig {
+    pub kms_signers: Option<Vec<String>>,
+    pub user_address: Option<String>,
+    pub gateway_chain_id: Option<u64>,
+    pub verifying_contract_address: Option<String>,
+    pub signature: Option<String>,
+    pub public_key: Option<String>,
+    pub private_key: Option<String>,
+    pub domain: Option<String>,
+    pub handle_contract_pairs: Option<Vec<CtHandleContractPair>>,
+    pub json_response: Option<String>,
+    pub verify_signatures: bool,
+}
+
+impl Default for ResponseConfig {
+    fn default() -> Self {
+        Self {
+            kms_signers: None,
+            user_address: None,
+            gateway_chain_id: None,
+            verifying_contract_address: None,
+            signature: None,
+            public_key: None,
+            private_key: None,
+            domain: Some("Decryption".to_string()),
+            handle_contract_pairs: None,
+            json_response: None,
+            verify_signatures: false,
+        }
+    }
+}
